+++ packages/add-changeset/src/index.ts
@@ -1,91 +1,7 @@
-import { Changeset, Release, VersionType } from "@changesets/types";
 import { getInput, setOutput, setFailed, getMultilineInput } from "@actions/core";
-import glob, { IOptions } from "glob";
-import { existsSync } from "fs";
-import { readFile } from "fs/promises";
-import write from "@changesets/write";
-import path from "path";
+import { addChangeset } from "./addChangeset";
 
-type AddChangesetParameters = {
-    filter?: RegExp;
-    ignore?: Array<string>;
-    type?: VersionType;
-    summary: string;
-    cwd?: string;
-};
-
-const globAsync = (pattern: string, options: IOptions): Promise<string[]> => {
-    return new Promise<string[]>((resolve, reject) => {
-        glob(pattern, options, (error, matches) => {
-            if (error) {
-                reject(error);
-            } else {
-                resolve(matches);
-            }
-        });
-    });
-};
-
-type PackageJson = {
-    name: string;
-};
-
-const loadProjectPackageJson = async (packageJsonPath: string): Promise<PackageJson> => {
-    if (!existsSync(packageJsonPath)) {
-        throw new Error(`Could not locate package.json in ${packageJsonPath}`);
-    }
-
-    const contents = await readFile(packageJsonPath);
-
-    return JSON.parse(contents.toString());
-};
-
-type CreatedChangeset = {
-    name: string;
-    path: string;
-};
-
-async function addChangeset({ filter, ignore, type, summary, cwd }: AddChangesetParameters): Promise<CreatedChangeset> {
-    const packageJsonFilePaths = await globAsync("**/package.json", { absolute: true, ignore: ignore });
-
-    const releases: Release[] = [];
-
-    for (const packageJsonPath of packageJsonFilePaths) {
-        const packageJson = await loadProjectPackageJson(packageJsonPath);
-
-        let addPackage = true;
-
-        if (filter !== undefined && !filter.test(packageJson.name)) {
-            addPackage = false;
-        }
-
-        if (addPackage) {
-            releases.push({
-                name: packageJson.name,
-                type: type || "patch",
-            });
-        }
-    }
-
-    const changeset: Changeset = {
-        summary,
-        releases,
-    };
-
-    console.log(`Writing changeset '${JSON.stringify(changeset)}'`);
-
-    const workingDirectory = cwd || ".";
-    const changesetDirectory = path.join(workingDirectory, ".changeset");
-
-    const changesetName = await write(changeset, cwd || ".");
-
-    const changesetPath = path.join(changesetDirectory, changesetName);
-
-    console.log(`Changeset '${changesetName}' created at '${changesetPath}'`);
-
-    return { name: changesetName, path: changesetPath };
-}
-
+// todo: convert to use context when this is available
 async function addChangesetAction() {
     try {
         const filter = getInput("filter");
